Pull out sign-up related functionality into a separate module
