+++ src/modules/signup/components/CreateOrganizationForm.js
@@ -4,7 +4,7 @@ import { withStyles } from "@material-ui/core/styles";
 import { TextField } from "formik-material-ui";
 import Typography from "@material-ui/core/Typography";
 
-import ButtonWithProgress from "./ButtonWithProgress";
+import ButtonWithProgress from "components/ButtonWithProgress";
 
 const styles = theme => ({
   submit: {
@@ -12,23 +12,15 @@ const styles = theme => ({
   }
 });
 
-const SignUpForm = ({ classes, onSubmit, submitButtonText }) => (
+const CreateOrganizationForm = ({ classes, onSubmit }) => (
   <Formik
     initialValues={{
-      name: "",
-      email: "",
-      password: ""
+      name: ""
     }}
     validate={values => {
       let errors = {};
       if (!values.name) {
-        errors.name = "Enter your name.";
-      }
-      if (!values.email) {
-        errors.email = "Enter your email.";
-      }
-      if (!values.password) {
-        errors.password = "Enter a password.";
+        errors.name = "Enter a name.";
       }
       return errors;
     }}
@@ -38,25 +30,9 @@ const SignUpForm = ({ classes, onSubmit, submitButtonText }) => (
         {/* TODO: style errors */}
         {errors.form ? <Typography>{errors.form}</Typography> : null}
         <Field
-          type="name"
-          label="Enter your name"
+          type="text"
+          label="What’s the name of your organization?"
           name="name"
-          component={TextField}
-          fullWidth
-          margin="normal"
-        />
-        <Field
-          type="email"
-          label="What is your email?"
-          name="email"
-          component={TextField}
-          fullWidth
-          margin="normal"
-        />
-        <Field
-          type="password"
-          label="Choose a password"
-          name="password"
           component={TextField}
           fullWidth
           margin="normal"
@@ -68,15 +44,11 @@ const SignUpForm = ({ classes, onSubmit, submitButtonText }) => (
           loading={isSubmitting}
           className={classes.submit}
         >
-          {submitButtonText}
+          Continue
         </ButtonWithProgress>
       </form>
     )}
   />
 );
 
-SignUpForm.defaultProps = {
-  submitButtonText: "Sign Up"
-};
-
-export default withStyles(styles)(SignUpForm);
+export default withStyles(styles)(CreateOrganizationForm);
