+++ src/modules/signup/containers/InviteLink.js
@@ -7,10 +7,10 @@ import TextField from "@material-ui/core/TextField";
 import Button from "@material-ui/core/Button";
 import Snackbar from "@material-ui/core/Snackbar";
 
-import { getOrAddInvite, createInviteLink } from "queries/invites";
 import { waitForProfile, handleError } from "utils";
+import Progress from "components/Progress";
 
-import Progress from "./Progress.js";
+import { getOrAddInvite, createInviteLink } from "../queries/invites";
 
 class InviteLink extends React.Component {
   state = {
