+++ src/modules/signup/pages/CreateOrganizationFlow.js
@@ -2,13 +2,13 @@ import React from "react";
 import Typography from "@material-ui/core/Typography";
 import Button from "@material-ui/core/Button";
 
-import { addOrganization } from "queries/organizations";
 import { createUserAndProfile } from "auth";
 import { handleError } from "utils";
 
-import CreateOrganizationForm from "./CreateOrganizationForm";
-import SignUpForm from "./SignUpForm";
-import InviteLink from "./InviteLink";
+import { addOrganization } from "../queries/organizations";
+import CreateOrganizationForm from "../components/CreateOrganizationForm";
+import SignUpForm from "../components/SignUpForm";
+import InviteLink from "../containers/InviteLink";
 
 const OrganizationStep = ({ onDone }) => {
   return (
