+++ src/modules/signup/pages/InvitePage.js
@@ -2,7 +2,8 @@ import React from "react";
 
 import AppFrame from "components/AppFrame";
 import Page from "components/Page";
-import InviteLink from "components/InviteLink";
+
+import InviteLink from "../containers/InviteLink";
 
 const InvitePage = () => (
   <AppFrame title="Invite people">
